added reward scale for DoodleJump-v0
When the magnitudes of the rewards are very large, PPO+Adam has a lot of trouble training the value function. Why? Because the value function's magnitude is on the order of ~200, and Adam only allows us to change each weight by ~1e-3 at each update. I've found that the value function tends to collapse to a single value when the rewards are very large.