change mouse click/move order
The agent outputs a probability distribution of movements, so it doesn't know where the mouse will end up in the next timestep. Thus, it makes more sense to have the agent to decide whether or not to click the location it *does* know about, namely the location as seen in the previous observation.