DoodleJump-v0 script
This script uses basic pixel-based search to find objects (platforms and the character) on the screen. It is stochastic, but always presses a single key at a time and never presses the shoot key.

This game has been giving me trouble since I started trying to train agents on it. I think this script might already beat the best agents I ever trained (on average). With this script as a starting point, I can hopefully get more out of my DoodleJump agents.